refactor: finalize project structure with updated packages, mappers, and DTOs
