+++ src/main/java/SpringBootApp/App/Project/payloads/ProjectRequestDto.java
@@ -1,0 +1,13 @@
+package SpringBootApp.App.Project.payloads;
+
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class ProjectRequestDto {
+    private String title;
+
+}
