+++ src/main/java/SpringBootApp/App/Project/repository/ProjectRepository.java
@@ -1,0 +1,10 @@
+package SpringBootApp.App.Project.repository;
+
+import SpringBootApp.App.Project.entities.ProjectEntity;
+import com.sparktechcode.springjpasearch.repositories.SparkJpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProjectRepository extends SparkJpaRepository<String, ProjectEntity> {
+
+}
