+++ src/main/java/SpringBootApp/App/Tasks/entities/Task.java
@@ -1,0 +1,4 @@
+package SpringBootApp.App.Tasks.entities;
+
+public interface Task {
+}
