+++ src/main/java/SpringBootApp/App/Tasks/payloads/TaskRequestDto.java
@@ -1,0 +1,17 @@
+package SpringBootApp.App.Tasks.payloads;
+
+import SpringBootApp.App.general.Status;
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class TaskRequestDto {
+
+    private String title;
+    private String description;
+    private Status status;
+    private String userId;
+
+}
