+++ src/main/java/SpringBootApp/App/Tasks/payloads/TaskUserDto.java
@@ -1,0 +1,14 @@
+package SpringBootApp.App.Tasks.payloads;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class TaskUserDto {
+
+    private String taskId;
+    private String userId;
+}
