+++ src/main/java/SpringBootApp/App/User/entities/UserEntity.java
@@ -1,0 +1,33 @@
+package SpringBootApp.App.User.entities;
+
+import SpringBootApp.App.Project.entities.ProjectEntity;
+import com.sparktechcode.springcrud.entities.AutoIdBaseEntity;
+import jakarta.persistence.*;
+import lombok.*;
+import java.util.ArrayList;
+import java.util.List;
+
+@Getter
+@Setter
+@Entity
+@Table(name="users")
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString(onlyExplicitlyIncluded = true, callSuper = true)
+@EqualsAndHashCode(onlyExplicitlyIncluded = true, callSuper = false)
+public class UserEntity extends AutoIdBaseEntity<String> {
+
+    private String firstName;
+    private String lastName;
+    private String email;
+    private String password;
+
+    @ManyToMany(fetch = FetchType.LAZY)
+    @JoinTable(
+            name="user_project",
+            joinColumns = @JoinColumn(name="user_id"),
+            inverseJoinColumns = @JoinColumn(name = "project_id")
+    )
+    private List<ProjectEntity> projects=new ArrayList<>();
+
+}
