+++ src/main/java/SpringBootApp/App/User/payloads/UserProjectRequestDto.java
@@ -1,0 +1,13 @@
+package SpringBootApp.App.User.payloads;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserProjectRequestDto {
+
+    private String projectId;
+}
