+++ src/main/java/SpringBootApp/App/User/payloads/UserRequestDto.java
@@ -1,0 +1,16 @@
+package SpringBootApp.App.User.payloads;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserRequestDto {
+
+    private String firstName;
+    private String lastName;
+    private String email;
+    private String password;
+
+}
