--- src/main/java/SpringBootApp/App/entities/Status.java
@@ -1,8 +1,0 @@
-package SpringBootApp.App.entities;
-
-public enum Status {
-    TO_DO,
-    IN_PROGRESS,
-    IN_REVIEW,
-    DONE
-}
