--- src/main/java/SpringBootApp/App/entities/UserEntity.java
@@ -1,38 +1,0 @@
-package SpringBootApp.App.entities;
-
-import com.sparktechcode.springjpasearch.entities.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-
-import org.hibernate.annotations.GenericGenerator;
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Table(name="users")
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@ToString(onlyExplicitlyIncluded = true, callSuper = true)
-@EqualsAndHashCode(onlyExplicitlyIncluded = true, callSuper = false)
-public class UserEntity implements BaseEntity<String> {
-
-    @Id
-    @GeneratedValue(generator = "uuid")
-    @GenericGenerator(name="uuid", strategy = "uuid")
-    private String Id;
-
-    private String firstName;
-    private String lastName;
-    private String email;
-    private String password;
-
-    @ManyToMany
-    @JoinTable(
-            name="user_project",
-            joinColumns = @JoinColumn(name="user_id"),
-            inverseJoinColumns = @JoinColumn(name = "project_id")
-    )
-    private List<ProjectEntity> projects=new ArrayList<>();
-
-}
