+++ src/main/java/SpringBootApp/App/general/AuditDto.java
@@ -1,0 +1,13 @@
+package SpringBootApp.App.general;
+
+import lombok.Getter;
+import lombok.Setter;
+import java.time.Instant;
+
+@Getter
+@Setter
+public class AuditDto {
+
+    private Instant created;
+    private Instant updated;
+}
