+++ src/main/java/SpringBootApp/App/general/Status.java
@@ -1,0 +1,8 @@
+package SpringBootApp.App.general;
+
+public enum Status {
+    TO_DO,
+    IN_PROGRESS,
+    IN_REVIEW,
+    DONE
+}
