--- src/main/java/SpringBootApp/App/payloads/ProjectRequestDto.java
@@ -1,14 +1,0 @@
-package SpringBootApp.App.payloads;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ProjectRequestDto {
-    private String title;
-
-}
