--- src/main/java/SpringBootApp/App/payloads/TaskUserDto.java
@@ -1,14 +1,0 @@
-package SpringBootApp.App.payloads;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class TaskUserDto {
-
-    private String taskId;
-    private String userId;
-}
