--- src/main/java/SpringBootApp/App/payloads/UserProjectRequestDto.java
@@ -1,13 +1,0 @@
-package SpringBootApp.App.payloads;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserProjectRequestDto {
-
-    private String projectId;
-}
