--- src/main/java/SpringBootApp/App/payloads/UserRequestDto.java
@@ -1,17 +1,0 @@
-package SpringBootApp.App.payloads;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserRequestDto {
-
-    private String firstName;
-    private String lastName;
-    private String email;
-    private String password;
-
-}
